bug fix: about interned symbol.
In previous implementation, first create an un-interned symbol, then
intern it into a symbol-table.

But I decided to change the interface, when the second argument of `voba_make_symbol`
is nil, it is created a global symbol, not an un-interned symbol.

so the bug fix just simply created an interned symbol instead.